Simplify EmojiBubbleChart props
The computed props makes sense in a world where we don't want
to re-render any time the input data changes; we gain very little
by trying to peel this computation out into a pre-process.

On a slow device, e.g. WP 8, doing this in render makes no change
to the total redraw time vs doing this in a computed prop, and the
abstraction is confusing.