Fix remaining Emoji bubble chart layout & animation issues
Internet Explorer was exhibiting interaction-blocking layout bugs, and
the other browsers could have used some polish; this PR finishes that
polishing. A more thorough explanation of what's going on in the CSS
is also provided as a comment in the sidecar SASS file, for future
clarity and debugging assistance.

Fixes #72

Internet Explorer screen capture:

![ie emoji pop-in fixed](https://cloud.githubusercontent.com/assets/442115/19929357/56bcc13c-a0d9-11e6-9b16-3fbae58a8ae3.gif)